Update testing folders
