+++ src/__tests__/Header.test.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import { BrowserRouter } from 'react-router-dom';
+import renderer from 'react-test-renderer';
+
+import Header from '../components/Header';
+
+test('it should render correctly', () => {
+  const tree = renderer.create(<BrowserRouter><Header /></BrowserRouter>).toJSON();
+  expect(tree).toMatchSnapshot();
+});
