+++ src/__tests__/Quote.test.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import renderer from 'react-test-renderer';
+import Quote from '../components/Quote';
+
+test('it should render correctly', () => {
+  const tree = renderer.create(<Quote />).toJSON();
+  expect(tree).toMatchSnapshot();
+});
