+++ src/__tests__/calculate.test.js
@@ -1,0 +1,12 @@
+// import { isNumber, calculate } from '../logic/calculate';
+
+// describe('isNumber', () => {
+// console.log(isNumber)
+// //   test('Shuold return true if the input is a number', () => {
+// //     expect(isNumber('5')).toBeTruthy();
+// //   });
+
+// //   test('Should return false if  the input value is not a number', () => {
+// //     expect(isNumber('xxx')).toBeFalsy();
+// //   });
+// });
