--- src/components/Header.test.js
@@ -1,10 +1,0 @@
-import React from 'react';
-import { BrowserRouter } from 'react-router-dom';
-import renderer from 'react-test-renderer';
-
-import Header from './Header'
-
-test ('it should render correctly', () => {
-    const tree = renderer.create(<BrowserRouter><Header /></BrowserRouter>).toJSON();
-    expect(tree).toMatchSnapshot();
-})
