--- src/components/Quote.test.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import renderer from 'react-test-renderer';
-import Quote from './Quote';
-
-test ('it should render correctly', () => {
-    const tree = renderer.create(<Quote />).toJSON();
-    expect(tree).toMatchSnapshot();
-})
