+++ src/logic/calculate.js
@@ -1,6 +1,6 @@
 import operate from './operate';
 
-export function isNumber(item) {
+function isNumber(item) {
   return !!item.match(/[0-9]+/);
 }
 
@@ -13,7 +13,7 @@ export function isNumber(item) {
  *   next:String       the next number to be operated on with the total
  *   operation:String  +, -, etc.
  */
-export default function calculate(obj, buttonName) {
+function calculate(obj, buttonName) {
   if (buttonName === 'AC') {
     return {
       total: null,
@@ -118,3 +118,5 @@ export default function calculate(obj, buttonName) {
     operation: buttonName,
   };
 }
+
+export { isNumber, calculate };
