--- src/logic/calculate.test.js
@@ -1,12 +1,0 @@
-// import { isNumber } from "./calculate";
-
-// describe('isNumber', () => {
-//     test('Shuold return true if the input is a number', () => {
-//         expect(isNumber('5')).toBeTruthy();
-//     });
-
-//     test('Should return false if  the input value is not a number', () => {
-//         expect(isNumber('xxx')).toBeFalsy();
-//     });
-// })
-
